play header issue fixes
